requirements.txt
pyswisseph
fastapi
uvicorn
